Multiple saves of records not beneficial for http
We must introduce an id field for Record object to be able to consistently update the proper fields during the run. While have the status in lable is interesting as it allows to track down lively the process while running with sumatra (which was not possible before: we had to wait until the end) it is more sound to have the status label in the status field. We will be working on that later and have a Chat with Dr. Davison.